% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.handle_data_parameter}
\alias{.handle_data_parameter}
\title{Take any kind of data and extract the matrix}
\usage{
.handle_data_parameter(data, on_disk, allow_sparse = TRUE)
}
\description{
Adapted from glmGamPoi:::handle_data_parameter
}
\keyword{internal}
